Use application context when clear image with Glide
Glide.with(Context) will check if the activity is destoryed when using
activity context. Since we are manually clearing the image view
drawable, instead of loading drawables, we should use the application
context to avoid this check.
Also checked all the glide usage and moved the one in
ConferenceProfileViewHolder to use the util setContactBitmapAsync method.

Fixes: 185913036
Test: manually
Change-Id: Ib9de58e5ae98f95f764fa02b05fb78f2b67fabb3