Show HUN for self managed calls
Show HUN for self managed calls so voip apps don't have to do this. On
another hand, no matter if the voip app shows the HUN for its
own ringing call, we will display the HUN for incoming self managed
calls.

VoIP app should pass the component name of their incall activity to the
call details extras if the incall activity is not the launch activity.

Bug: 241035588
Test: manually

Change-Id: I663f245e6ddc3ee0ddf3794c5d364eae9cd20bdb