Proactively set apps to compile with Java 8
Similar to ag/31409817, now that we're dropping from Soong again we have to make sure the Soong builds are compatible with sc_car. Sc_car only support Java 8, so each of the apps that are dropped to it need to compile to Java 8. The default Java version is now newer than 8, so we need to manually set it.

Test: Presubmit
Bug: b/391876191
Change-Id: I000cb553fddcaf7ab24001c2d4047834e31dbfc8