Update Dialer to use version catalog
This ensures that Dialer uses the same dependency versions as the rest of the build.

This required backing down the androidx-test:runner version to 1.6.1 to put it in sync with androidx-test:rules:1.6.1, which had a hard dependency on the runner. There weren't any release notes for version 1.6.2 so I assume there were no major changes.

Also needed to update car-ui-lib:testing since it was also specifying specific, out of date versions.

Bug: b/389982145
Test: ./busytown/run-host-unit-tests.sh
Change-Id: Ia7598807e775468855a1af181355e7684089e1f2